Changing expiration type from `Instant` to `SystemTime`
# Situation
Currently, the expiration time is captured as an `Instant`. This is fine for comparing against other instants. However, it doesn't work well when trying to work with other time types.

# Target
Change the expiration time to a type that can be more easily be used across various time types.

# Proposal
Move from using `Instant` to `SystemTime`.