Move from `SystemTime` to `Instant`
# Situation
Currently, we use `SystemTime` for our expiration time. However, if the host system time is off or tampered with, it can cause the cache to purge when it shouldn't.

# Target
Ensure that our expiration time is not subject to host inaccuracies or tampering.

# Proposal
Move to using `Instant`, which does not reference the host system time.